securethenews/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file /code/requirements.txt /code/requirements.in
#
-e git+https://github.com/freedomofpress/django-logging.git@51f4fd4b02fafc3d6d9cddb5658247497404df7e#egg=django-logging-json
asn1crypto==0.22.0        # via cryptography
beautifulsoup4==4.6.0     # via pytablereader, wagtail
boto3==1.7.48
botocore==1.10.48         # via boto3, s3transfer
cachetools==2.1.0         # via google-auth
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==1.9         # via pyopenssl, sslyze
dataproperty==0.25.6      # via pytablereader, pytablewriter, simplesqlite
django-analytical==2.2.2
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-mailgun==0.9.1
django-modelcluster==3.1  # via wagtail
django-storages[boto3,google]==1.6.6
django-taggit==0.22.1     # via wagtail
django-treebeard==4.1.2   # via wagtail
django==1.11.15
djangorestframework==3.6.4  # via wagtail
docopt==0.6.2             # via pshtt
docutils==0.14            # via botocore
dominate==2.3.1           # via pytablewriter
elasticsearch==5.4.0
google-api-core==1.2.1    # via google-cloud-core, google-cloud-storage
google-auth==1.5.0        # via google-api-core
google-cloud-core==0.28.1  # via google-cloud-storage
google-cloud-storage==1.10.0  # via django-storages
google-resumable-media==0.3.1  # via google-cloud-storage
googleapis-common-protos==1.5.3  # via google-api-core
gunicorn==19.7.1
html5lib==0.999999999     # via wagtail
idna==2.6                 # via cryptography, requests
jmespath==0.9.3           # via boto3, botocore
jsonschema==2.6.0         # via pytablereader
logbook==1.1.0            # via dataproperty, pytablereader, pytablewriter, simplesqlite
markdown2==2.3.5
mbstrdecoder==0.2.2       # via pytablereader, pytablewriter, simplesqlite, typepy
nassl==0.17.0             # via sslyze
olefile==0.44             # via pillow
path.py==10.4             # via pytablereader
pathvalidate==0.16.1      # via pytablereader, pytablewriter, simplesqlite
pillow==4.2.1             # via wagtail
protobuf==3.6.0           # via google-api-core, googleapis-common-protos
pshtt==0.2.3
psycopg2==2.7.3.1
publicsuffix==1.1.0       # via pshtt
pyasn1-modules==0.2.2     # via google-auth
pyasn1==0.4.3             # via pyasn1-modules, rsa
pycparser==2.18           # via cffi
pyopenssl==17.2.0         # via pshtt
pyparsing==2.2.0          # via pytablereader
pytablereader==0.13.3     # via pytablewriter, simplesqlite
pytablewriter==0.24.0     # via pshtt
python-dateutil==2.6.1    # via botocore, typepy
python-json-logger==0.1.8
pytz==2017.2              # via django, django-modelcluster, google-api-core, typepy
requests-cache==0.4.13    # via pshtt
requests==2.18.4          # via django-mailgun, google-api-core, pshtt, pytablereader, requests-cache, wagtail
rsa==3.4.2                # via google-auth
s3transfer==0.1.13        # via boto3
simplesqlite==0.15.0      # via pytablewriter
six==1.10.0               # via cryptography, django-mailgun, google-api-core, google-auth, google-resumable-media, html5lib, mbstrdecoder, protobuf, pyopenssl, pytablereader, pytablewriter, python-dateutil, simplesqlite, typepy, unittest-xml-reporting
sslyze==1.1.4             # via pshtt
tls-parser==1.1.0         # via sslyze
toml==0.9.2               # via pytablewriter
typepy==0.0.20            # via dataproperty, pytablereader, pytablewriter, simplesqlite
typing==3.6.2             # via nassl, sslyze, tls-parser
unidecode==0.4.21         # via wagtail
unittest-xml-reporting==2.1.1
urllib3==1.22             # via elasticsearch, requests
wagtail==1.12.1
wagtailmenus==2.4
webencodings==0.5.1       # via html5lib
wget==3.2                 # via pshtt
willow==0.4               # via wagtail
xlrd==1.1.0               # via pytablereader
xlsxwriter==0.9.9         # via pytablewriter
xlwt==1.3.0               # via pytablewriter
